# app/api/openapi.json
{
  "info": {
    "version": "0.0.1",
    "title": "A swagger API",
    "description": "powered by Flasgger",
    "termsOfService": "/tos"
  },
  "paths": {
...
    "/api/settings/default/brightness": {
      "get": {
        "summary": "Get or update default brightness setting.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update default brightness setting.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/default/mode": {
      "get": {
        "summary": "Get or update default operating mode setting.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update default operating mode setting.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/mode/intervals/flash": {
      "get": {
        "summary": "Get or update flash interval setting.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update flash interval setting.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/mode/intervals/wave": {
      "get": {
        "summary": "Get or update wave interval setting.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update wave interval setting.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/mode/intervals/scatter": {
      "get": {
        "summary": "Get or update scatter interval setting.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update scatter interval setting.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/status/{setting_key}": {
      "get": {
        "summary": "Get or update a status setting by key.",
        "parameters": [
          {
            "name": "setting_key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique setting_key of the status (e.g. available_status)"
          }
        ],
        "tags": [
          "settings_api"
        ]
      },
      "put": {
        "summary": "Get or update a status setting by key.",
        "parameters": [
          {
            "name": "setting_key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique setting_key of the status (e.g. available_status)"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "value"
              ],
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Get or update a status setting by key.",
        "parameters": [
          {
            "name": "setting_key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique setting_key of the status (e.g. available_status)"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "value"
              ],
              "properties": {
                "value": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      },
      "delete": {
        "summary": "Delete a status setting by key.",
        "responses": {
          "200": {
            "description": "Status deleted successfully",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "No status found with that key",
            "schema": {
              "type": "string"
            }
          }
        },
        "parameters": [
          {
            "name": "setting_key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Unique setting_key of the status to delete"
          }
        ],
        "produces": [
          "text/plain"
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/status": {
      "get": {
        "summary": "GET all statuses or POST a new status to the Statuses group.",
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "GET all statuses or POST a new status to the Statuses group.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "setting_key",
                "value"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "setting_key": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
...
    "/api/settings": {
      "get": {
        "summary": "Fetch all settings groups with nested settings matching your structured JSON output.",
        "responses": {
          "200": {
            "description": "List of groups with nested settings",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "settings_api"
        ]
      },
      "post": {
        "summary": "Create a new individual Setting.",
        "responses": {
          "201": {
            "description": "Setting created successfully"
          },
          "400": {
            "description": "Missing required fields or invalid value_type"
          },
          "404": {
            "description": "Group ID not found"
          },
          "409": {
            "description": "Setting key already exists"
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name",
                "setting_key",
                "value",
                "settings_group_id"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "setting_key": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                },
                "settings_group_id": {
                  "type": "integer"
                },
                "value_type": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/settings/{key}": {
      "get": {
        "summary": "Fetch any single setting dynamically by key using to_dict().",
        "responses": {
          "200": {
            "description": "Setting found",
            "schema": {
              "type": "object"
            }
          },
          "404": {
            "description": "Setting not found"
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Setting key identifier"
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/groups/{group_id}": {
      "delete": {
        "summary": "Delete a setting group (cascades to associated settings).",
        "responses": {
          "200": {
            "description": "Group deleted successfully"
          },
          "404": {
            "description": "Group not found"
          }
        },
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Group ID to delete"
          }
        ],
        "tags": [
          "settings_api"
        ]
      }
    },
    "/api/groups": {
      "post": {
        "summary": "Create a new Settings Group.",
        "responses": {
          "201": {
            "description": "Group created successfully"
          },
          "400": {
            "description": "Field name is required"
          },
          "409": {
            "description": "Group already exists"
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "tags": [
          "settings_api"
        ]
      }
    }
  },
...
  "swagger": "2.0"
}